Changes to broadcast to support .<, .==, etc for indices
Broadcast looks like it needs rewriting for jbfunctions. This starts on the process for these specific operations, but leaves the rest of the problem untouched.

One issue here is that I'm implicitly assuming that `parent(P)` returns the "inner" type requested by `allocate_for`. I haven't had to make that assumption yet.